+++ changelog.md
@@ -3,9 +3,9 @@
 This project adheres to semantic versioning and all major changes will
 be recorded in this file.
 
-## [unreleased]
+## [0.4.0] 2023-07-16
 
-- Add type MakeGet
+- Add types MakeGet and MakeSet
 - Fix field type for non ast.Ident expressions
 
 ## [0.3.0] 2023-07-16